Added gitignore, added solution/project files, fixed file structure, added xunit dependency and added namespaces.
