+++ C#/TollCalculator/TollCalculator.Lib/Car.cs
@@ -1,10 +1,6 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
-namespace TollFeeCalculator
+namespace TollCalculator.Lib
 {
     public class Car : Vehicle
     {
