+++ C#/TollCalculator/TollCalculator.Lib/Motorbike.cs
@@ -1,0 +1,10 @@
+﻿namespace TollCalculator.Lib
+{
+    public class Motorbike : Vehicle
+    {
+        public string GetVehicleType()
+        {
+            return "Motorbike";
+        }
+    }
+}
