+++ C#/TollCalculator/TollCalculator.Lib/Vehicle.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace TollCalculator.Lib
+{
+    public interface Vehicle
+    {
+        String GetVehicleType();
+    }
+}
