+++ C#/TollCalculator/TollCalculator.sln
@@ -1,0 +1,48 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+VisualStudioVersion = 15.0.26124.0
+MinimumVisualStudioVersion = 15.0.26124.0
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TollCalculator.Lib", "TollCalculator.Lib\TollCalculator.Lib.csproj", "{D7A202F2-F9B5-4E9D-B0B1-997E9BCB0B35}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TollCalculator.Tests", "TollCalculator.Tests\TollCalculator.Tests.csproj", "{C0DECB7D-C44F-48DE-BC18-D5AA0C201D0A}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|Any CPU = Release|Any CPU
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{D7A202F2-F9B5-4E9D-B0B1-997E9BCB0B35}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D7A202F2-F9B5-4E9D-B0B1-997E9BCB0B35}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D7A202F2-F9B5-4E9D-B0B1-997E9BCB0B35}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{D7A202F2-F9B5-4E9D-B0B1-997E9BCB0B35}.Debug|x64.Build.0 = Debug|Any CPU
+		{D7A202F2-F9B5-4E9D-B0B1-997E9BCB0B35}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{D7A202F2-F9B5-4E9D-B0B1-997E9BCB0B35}.Debug|x86.Build.0 = Debug|Any CPU
+		{D7A202F2-F9B5-4E9D-B0B1-997E9BCB0B35}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D7A202F2-F9B5-4E9D-B0B1-997E9BCB0B35}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D7A202F2-F9B5-4E9D-B0B1-997E9BCB0B35}.Release|x64.ActiveCfg = Release|Any CPU
+		{D7A202F2-F9B5-4E9D-B0B1-997E9BCB0B35}.Release|x64.Build.0 = Release|Any CPU
+		{D7A202F2-F9B5-4E9D-B0B1-997E9BCB0B35}.Release|x86.ActiveCfg = Release|Any CPU
+		{D7A202F2-F9B5-4E9D-B0B1-997E9BCB0B35}.Release|x86.Build.0 = Release|Any CPU
+		{C0DECB7D-C44F-48DE-BC18-D5AA0C201D0A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C0DECB7D-C44F-48DE-BC18-D5AA0C201D0A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C0DECB7D-C44F-48DE-BC18-D5AA0C201D0A}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{C0DECB7D-C44F-48DE-BC18-D5AA0C201D0A}.Debug|x64.Build.0 = Debug|Any CPU
+		{C0DECB7D-C44F-48DE-BC18-D5AA0C201D0A}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{C0DECB7D-C44F-48DE-BC18-D5AA0C201D0A}.Debug|x86.Build.0 = Debug|Any CPU
+		{C0DECB7D-C44F-48DE-BC18-D5AA0C201D0A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C0DECB7D-C44F-48DE-BC18-D5AA0C201D0A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C0DECB7D-C44F-48DE-BC18-D5AA0C201D0A}.Release|x64.ActiveCfg = Release|Any CPU
+		{C0DECB7D-C44F-48DE-BC18-D5AA0C201D0A}.Release|x64.Build.0 = Release|Any CPU
+		{C0DECB7D-C44F-48DE-BC18-D5AA0C201D0A}.Release|x86.ActiveCfg = Release|Any CPU
+		{C0DECB7D-C44F-48DE-BC18-D5AA0C201D0A}.Release|x86.Build.0 = Release|Any CPU
+	EndGlobalSection
+EndGlobal
